+++ test.cpp
@@ -1,6 +1,0 @@
-#include <iostream>
-
-int main() {
-    std::cout << "This is first test!";
-    return 0;
-}