+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.1.1 - 15/01/21
 - Added to-do list
 - Changed repo name and local project name from 'CPP-Project' to 'Agent_Survival_Simulation'
+- Added BasicAgent and BasicEnvironment classes to 'basic_simulation.cpp'
+- Added Obstacle and Energy structs to 'basic_simulation.cpp'
 
 ### 0.1.0 - 13/01/21
 - Initial commit